refactor(fbuild-cli): split mcp.rs into submodules to satisfy LOC gate
