+++ crates/fbuild-cli/src/mcp/jsonrpc.rs
@@ -1,0 +1,83 @@
+//! JSON-RPC 2.0 envelope types used by the MCP stdio server.
+
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+#[derive(Debug, Deserialize)]
+pub(super) struct JsonRpcRequest {
+    pub(super) jsonrpc: String,
+    pub(super) id: Option<Value>,
+    pub(super) method: String,
+    #[serde(default)]
+    pub(super) params: Option<Value>,
+}
+
+#[derive(Debug, Serialize)]
+pub(super) struct JsonRpcResponse {
+    jsonrpc: String,
+    id: Value,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    result: Option<Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    error: Option<JsonRpcError>,
+}
+
+#[derive(Debug, Serialize)]
+pub(super) struct JsonRpcError {
+    code: i32,
+    message: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    data: Option<Value>,
+}
+
+impl JsonRpcResponse {
+    pub(super) fn success(id: Value, result: Value) -> Self {
+        Self {
+            jsonrpc: "2.0".to_string(),
+            id,
+            result: Some(result),
+            error: None,
+        }
+    }
+
+    pub(super) fn error(id: Value, code: i32, message: String) -> Self {
+        Self {
+            jsonrpc: "2.0".to_string(),
+            id,
+            result: None,
+            error: Some(JsonRpcError {
+                code,
+                message,
+                data: None,
+            }),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn json_rpc_response_serialization() {
+        let resp =
+            JsonRpcResponse::success(Value::Number(1.into()), serde_json::json!({"key": "value"}));
+        let json = serde_json::to_string(&resp).unwrap();
+        assert!(json.contains("\"jsonrpc\":\"2.0\""));
+        assert!(json.contains("\"result\""));
+        assert!(!json.contains("\"error\""));
+    }
+
+    #[test]
+    fn json_rpc_error_serialization() {
+        let resp = JsonRpcResponse::error(
+            Value::Number(2.into()),
+            -32601,
+            "Method not found".to_string(),
+        );
+        let json = serde_json::to_string(&resp).unwrap();
+        assert!(json.contains("\"error\""));
+        assert!(json.contains("-32601"));
+        assert!(!json.contains("\"result\""));
+    }
+}
